só vai remover o pID referente ao arquivo em execução

+++ robo_monitor.py
@@ -110,15 +110,20 @@ while True:      #Fará looping infinito, buscando novos arquivos nas pastas - s
             try:
                 fileEpid = 0
                 for fileEpid in glob.glob("{}\\*.pid".format(localPid)):
-                    pId = fileEpid.split("__")
-                    pId = int (pId[-1].replace(".pid", ""))
-                    if (not(rf.checkPID(pId))): #se pID não está em execução. remover arquivos
-                        try:
-                            os.remove(fileEpid)
-                            print("Removido o PID: {}".format(pId))
-                            os.remove(infoLog)
-                        except:
-                            pass
+
+                    pId = fileEpid.split("\\")[-1]
+                    pId = pId.split("__")
+
+                    print(file.split('.')[0])
+                    if (pId[0] == file.split('.')[0]): #só vai remover o pID referente ao arquivo em execução
+                        pId = int (pId[-1].replace(".pid", ""))
+                        if (not(rf.checkPID(pId))): #se pID não está em execução. remover arquivos
+                            try:
+                                os.remove(fileEpid)
+                                print("Removido o PID: {}".format(pId))
+                                os.remove(infoLog)
+                            except:
+                                pass
                 if (not (fileEpid)):  #se FileEpid não existir, Remova o infoLog.
                     if (os.path.isfile(infoLog)): 
                         os.remove(infoLog)
